Ensure cursor div doesn't end up in clipboard
In a recent commit, the logic for displaying the cursor was moved into
the editor library, so that the host app does not have to draw the
cursor itself. This involves creating a DIV element with the class name
"corinthia-cursor". We already have a method called removeSpecial()
which gets rid of elements like this when copying content to the
clipboard, but it was not taking this case into account.