Fix invalid position exception in Scan.ts
When scanning through a document, e.g. during a spell checking or
find/replace session, we keep the current position in a variable. Until
now, this was not tracked, and thus it was possible for modifications to
the DOM tree (such as showing the cursor) to cause the position object
to become invalid. Now we always make sure the current position is
tracked.